# backend/docs/openapi.yaml
openapi: 3.0.1
info:
  title: Isa's Signs & Designs API
  description: Product catalog read API for storefront product listing and product
    detail pages.
  contact:
    name: Isa's Signs & Designs
    email: support@isasignsdesigns.local
  license:
    name: Internal Use
  version: v1
servers:
- url: http://localhost:8081
  description: Local development
tags:
- name: Products
  description: Catalog read endpoints
paths:
  /api/products:
    get:
      tags:
      - Products
      summary: List products
      description: Returns active products. Optionally filter by exact category value.
      operationId: list
      parameters:
      - name: category
        in: query
        description: Exact category filter (case-sensitive)
        required: false
        schema:
          type: string
        example: Home Decor
      responses:
        "200":
          description: Products returned
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ProductResponse"
              examples:
                list-products:
                  summary: Sample response
                  description: list-products
                  value:
                  - id: 23e9318b-272c-424e-99d0-671c90b8bcf6
                    slug: ramadan-lantern
                    name: Ramadan Lantern
                    description: Decorative lantern for Ramadan
                    category: Ramadan Decor
                    priceCents: 3499
                    currency: CAD
                    images:
                    - /assets/ramadan1.jpg
                    material: metal
                    isActive: true
  /api/products/{slug}:
    get:
      tags:
      - Products
      summary: Get product by slug
      description: Returns one active product by slug.
      operationId: bySlug
      parameters:
      - name: slug
        in: path
        description: Product slug
        required: true
        schema:
          type: string
        example: ramadan-lantern
      responses:
        "200":
          description: Product found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductResponse"
        "404":
          description: Product not found
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ProductResponse"
components:
  schemas:
    ProductResponse:
      type: object
      properties:
        id:
          type: string
          description: Product unique identifier
          format: uuid
          example: 23e9318b-272c-424e-99d0-671c90b8bcf6
        slug:
          type: string
          description: Unique URL slug for product pages
          example: ramadan-lantern
        name:
          type: string
          description: Display name
          example: Ramadan Lantern
        description:
          type: string
          description: Product description
          example: Decorative lantern for Ramadan
        category:
          type: string
          description: Category label used for storefront filtering
          example: Ramadan Decor
        priceCents:
          type: integer
          description: Price in minor units (cents)
          format: int64
          example: 3499
        currency:
          type: string
          description: Currency code
          example: CAD
        images:
          type: array
          description: Product image URLs
          items:
            type: string
            example: /assets/ramadan1.jpg
        material:
          type: string
          description: Primary material
          example: metal
        isActive:
          type: boolean
          description: Active flag used for soft-deactivation
          example: true
        createdAt:
          type: string
          description: Record creation timestamp
          format: date-time
          example: 2026-02-21T07:31:47.474497Z
        updatedAt:
          type: string
          description: Record last update timestamp
          format: date-time
          example: 2026-02-21T07:31:47.474497Z
      description: Product resource returned by the catalog API